babel: extract state, jsxUtils, attributes
