+++ src/babel/attributes.js
@@ -1,0 +1,333 @@
+exports.htmlAttributes = new Set([
+  // attributes
+  'abbr',
+  'accept',
+  'acceptCharset',
+  'accessKey',
+  'action',
+  'allowFullScreen',
+  'allowTransparency',
+  'alt',
+  'async',
+  'autoComplete',
+  'autoFocus',
+  'autoPlay',
+  'cellPadding',
+  'cellSpacing',
+  'challenge',
+  'charset',
+  'checked',
+  'cite',
+  'cols',
+  'colSpan',
+  'command',
+  'content',
+  'contentEditable',
+  'contextMenu',
+  'controls',
+  'coords',
+  'crossOrigin',
+  'data',
+  'dateTime',
+  'default',
+  'defer',
+  'dir',
+  'disabled',
+  'download',
+  'draggable',
+  'dropzone',
+  'encType',
+  'for',
+  'form',
+  'formAction',
+  'formEncType',
+  'formMethod',
+  'formNoValidate',
+  'formTarget',
+  'frameBorder',
+  'headers',
+  'height',
+  'hidden',
+  'high',
+  'href',
+  'hrefLang',
+  'htmlFor',
+  'httpEquiv',
+  'icon',
+  'id',
+  'inputMode',
+  'isMap',
+  'itemId',
+  'itemProp',
+  'itemRef',
+  'itemScope',
+  'itemType',
+  'kind',
+  'label',
+  'lang',
+  'list',
+  'loop',
+  'manifest',
+  'max',
+  'maxLength',
+  'media',
+  'mediaGroup',
+  'method',
+  'min',
+  'minLength',
+  'multiple',
+  'muted',
+  'name',
+  'noValidate',
+  'open',
+  'optimum',
+  'pattern',
+  'ping',
+  'placeholder',
+  'poster',
+  'preload',
+  'radioGroup',
+  'readOnly',
+  'rel',
+  'required',
+  'role',
+  'rows',
+  'rowSpan',
+  'sandbox',
+  'scope',
+  'scoped',
+  'scrolling',
+  'seamless',
+  'selected',
+  'shape',
+  'size',
+  'sizes',
+  'sortable',
+  'span',
+  'spellCheck',
+  'src',
+  'srcDoc',
+  'srcSet',
+  'start',
+  'step',
+  'style',
+  'tabIndex',
+  'target',
+  'title',
+  'translate',
+  'type',
+  'typeMustMatch',
+  'useMap',
+  'value',
+  'width',
+  'wmode',
+  'wrap',
+  // handlers
+  'onBlur',
+  'onChange',
+  'onClick',
+  'onContextMenu',
+  'onCopy',
+  'onCut',
+  'onDoubleClick',
+  'onDrag',
+  'onDragEnd',
+  'onDragEnter',
+  'onDragExit',
+  'onDragLeave',
+  'onDragOver',
+  'onDragStart',
+  'onDrop',
+  'onFocus',
+  'onInput',
+  'onKeyDown',
+  'onKeyPress',
+  'onKeyUp',
+  'onMouseDown',
+  'onMouseEnter',
+  'onMouseLeave',
+  'onMouseMove',
+  'onMouseOut',
+  'onMouseOver',
+  'onMouseUp',
+  'onPaste',
+  'onPointerDown',
+  'onPointerEnter',
+  'onPointerMove',
+  'onPointerUp',
+  'onPointerCancel',
+  'onPointerOut',
+  'onPointerLeave',
+  'onScroll',
+  'onSubmit',
+  'onTouchCancel',
+  'onTouchEnd',
+  'onTouchMove',
+  'onTouchStart',
+  'onWheel',
+])
+
+exports.htmlElements = new Set([
+  'a',
+  'abbr',
+  'address',
+  'area',
+  'article',
+  'aside',
+  'audio',
+  'b',
+  'base',
+  'bdi',
+  'bdo',
+  'blockquote',
+  'body',
+  'br',
+  'button',
+  'canvas',
+  'caption',
+  'cite',
+  'code',
+  'col',
+  'colgroup',
+  'data',
+  'datalist',
+  'dd',
+  'del',
+  'details',
+  'dfn',
+  'dialog',
+  'div',
+  'dl',
+  'dt',
+  'em',
+  'embed',
+  'fieldset',
+  'figcaption',
+  'figure',
+  'footer',
+  'form',
+  'h1',
+  'h2',
+  'h3',
+  'h4',
+  'h5',
+  'h6',
+  'head',
+  'header',
+  'hr',
+  'html',
+  'i',
+  'iframe',
+  'img',
+  'input',
+  'ins',
+  'kbd',
+  'keygen',
+  'label',
+  'legend',
+  'li',
+  'link',
+  'main',
+  'map',
+  'mark',
+  'math',
+  'menu',
+  'menuitem',
+  'meta',
+  'meter',
+  'nav',
+  'noscript',
+  'object',
+  'ol',
+  'optgroup',
+  'option',
+  'output',
+  'p',
+  'param',
+  'picture',
+  'pre',
+  'progress',
+  'q',
+  'rb',
+  'rp',
+  'rt',
+  'rtc',
+  'ruby',
+  's',
+  'samp',
+  'script',
+  'section',
+  'select',
+  'small',
+  'source',
+  'span',
+  'strong',
+  'style',
+  'sub',
+  'summary',
+  'sup',
+  'svg',
+  'table',
+  'tbody',
+  'td',
+  'template',
+  'textarea',
+  'tfoot',
+  'th',
+  'thead',
+  'time',
+  'title',
+  'tr',
+  'track',
+  'u',
+  'ul',
+  'variable',
+  'video',
+  'wbr',
+])
+
+exports.unitlessCssAttributes = new Set([
+  // Based on https://github.com/giuseppeg/style-sheet/blob/e71c119ecaccdb2e50be0759b45820b8cbf0c6df/src/data.js
+  'animationIterationCount',
+  'borderImageOutset',
+  'borderImageSlice',
+  'borderImageWidth',
+  'boxFlex',
+  'boxFlexGroup',
+  'boxOrdinalGroup',
+  'columnCount',
+  'columns',
+  'flex',
+  'flexGrow',
+  'flexPositive',
+  'flexShrink',
+  'flexNegative',
+  'flexOrder',
+  'gridRow',
+  'gridRowEnd',
+  'gridRowSpan',
+  'gridRowStart',
+  'gridColumn',
+  'gridColumnEnd',
+  'gridColumnSpan',
+  'gridColumnStart',
+  'fontWeight',
+  'lineClamp',
+  'lineHeight',
+  'opacity',
+  'order',
+  'orphans',
+  'tabSize',
+  'widows',
+  'zIndex',
+  'zoom',
+  // SVG
+  'fillOpacity',
+  'floodOpacity',
+  'stopOpacity',
+  'strokeDasharray',
+  'strokeDashoffset',
+  'strokeMiterlimit',
+  'strokeOpacity',
+  'strokeWidth',
+])
